pyproject.toml
[project]
name = "monarch-splitter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "monarchmoneycommunity>=1.3.0",
    "polars>=1.38.1",
    "pyyaml>=6.0.3",
]
